refact: move Flow-related structures to flow package
