+++ loadable/src/main/java/com/jeanbarrossilva/loadable/flow/ViewModel.extensions.kt
@@ -1,13 +1,13 @@
-package com.jeanbarrossilva.loadable.utils
+package com.jeanbarrossilva.loadable.flow
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jeanbarrossilva.loadable.Loadable
-import java.io.Serializable
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.emitAll
 import kotlinx.coroutines.launch
+import java.io.Serializable
 
 /**
  * Creates a [MutableStateFlow] with the given [initialValue] that gets all of [origin]'s values
@@ -19,7 +19,7 @@ import kotlinx.coroutines.launch
 @Suppress("KDocUnresolvedReference")
 fun <T : Serializable?> ViewModel.loadable(
     initialValue: T,
-    origin: suspend () -> Flow<Loadable<T>>
+    origin: suspend () -> Flow<Loadable<T>>,
 ): MutableStateFlow<Loadable<T>> {
     return loadable(initialValue).apply {
         viewModelScope.launch {
